requirements.txt
ogb==1.3.4
tensorboard==2.7.0
rdkit-pypi==2021.9.4
jupyter==1.0
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.7.1
tqdm==4.62.3
